requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pandas>=1.3.0
opencv-python>=4.5.0  # Optional: for advanced sea detection
tqdm>=4.62.0
json5>=0.9.0
streamlit>=1.28.0  # For interactive web interface 